Fix mismatch namespaces
