+++ src/Domain/Enums/ResourceSystem/AttendanceStatus.cs
@@ -1,4 +1,4 @@
-namespace DbApp.Domain.Enums;
+namespace DbApp.Domain.Enums.ResourceSystem;
 
 /// <summary>
 /// 员工考勤状态
