+++ src/Domain/Enums/ResourceSystem/EvaluationLevel.cs
@@ -1,4 +1,4 @@
-namespace DbApp.Domain.Enums;
+namespace DbApp.Domain.Enums.ResourceSystem;
 
 /// <summary>
 /// 员工绩效评级
