+++ src/Domain/Enums/ResourceSystem/MaintenanceType.cs
@@ -1,4 +1,4 @@
-namespace DbApp.Domain.Enums;
+namespace DbApp.Domain.Enums.ResourceSystem;
 
 /// <summary>
 /// 设施维护类型
