+++ src/Domain/Enums/ResourceSystem/TransactionType.cs
@@ -1,4 +1,4 @@
-namespace DbApp.Domain.Enums;
+namespace DbApp.Domain.Enums.ResourceSystem;
 
 /// <summary>
 /// 财务交易类型
